backend/requirements.txt
flask==3.0.2
flask-cors==5.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pillow==10.2.0
pandas==2.2.0
sqlalchemy==2.0.25
